mmc: queue: Match the data type of max_segments
JIRA: https://issues.redhat.com/browse/RHEL-41085

commit f6f607070aa6f0c730c5ca38e888c88f34bea7d6
Author: ChanWoo Lee <cw9316.lee@samsung.com>
Date:   Tue Aug 24 16:39:34 2021 +0900

    mmc: queue: Match the data type of max_segments

    Each function has a different data type for max_segments,
    Modify to match unsigned short(host->max_segs).
    * unsigned short          max_segs;       /* see blk_queue_max_segments */

    1) Return type : unsigned int
    static unsigned int mmc_get_max_segments(struct mmc_host *host)
    {
           return host->can_dma_map_merge ? MMC_DMA_MAP_MERGE_SEGMENTS :
                                             host->max_segs;
    }

    2) Parameter type : int
     mmc_alloc_sg(mmc_get_max_segments(host), gfp);
       -> static struct scatterlist *mmc_alloc_sg(int sg_len, gfp_t gfp)

    3) Parameter type : unsigned short
     blk_queue_max_segments(mq->queue, mmc_get_max_segments(host));
       -> void blk_queue_max_segments(struct request_queue *q,
                                            unsigned short max_segments)

    Acked-by: Coly Li <colyli@suse.de>
    Link: https://lore.kernel.org/r/20210824073934.19727-1-cw9316.lee@samsung.com

+++ drivers/mmc/core/queue.c
@@ -162,7 +162,7 @@ static void mmc_mq_recovery_handler(struct work_struct *work)
 	blk_mq_run_hw_queues(q, true);
 }
 
-static struct scatterlist *mmc_alloc_sg(int sg_len, gfp_t gfp)
+static struct scatterlist *mmc_alloc_sg(unsigned short sg_len, gfp_t gfp)
 {
 	struct scatterlist *sg;
 
@@ -191,7 +191,7 @@ static void mmc_queue_setup_discard(struct request_queue *q,
 		blk_queue_max_secure_erase_sectors(q, max_discard);
 }
 
-static unsigned int mmc_get_max_segments(struct mmc_host *host)
+static unsigned short mmc_get_max_segments(struct mmc_host *host)
 {
 	return host->can_dma_map_merge ? MMC_DMA_MAP_MERGE_SEGMENTS :
 					 host->max_segs;
